4.3.20 to work with 2/10/04 EQLive
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/trunk@489 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/spawn.cpp
@@ -81,8 +81,8 @@ QString print_item (uint16_t item)
   // assume no material name found
   const char *itemStr = NULL;
 
-  uint8_t itemLo = item & 0x00ff;
-  uint8_t itemHi = (item & 0xff00) >> 8;
+  uint32_t itemLo = item & 0x00ff;
+  uint32_t itemHi = (item & 0xff00) >> 8;
 
   if (itemHi == 0x00)
   {
